requirements.txt
fastapi==0.127.1
uvicorn[standard]==0.40.0
pyodbc==5.3.0
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
python-multipart==0.0.21
mido==1.3.3
music21==9.9.1
google-cloud-secret-manager==2.26.0
